requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
scikit-learn
python-dotenv
openai